backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.120.1
# h11 pinned removed to avoid conflicts between uvicorn/httpcore/httpx
idna==3.11
mysql-connector-python==9.5.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0


SQLAlchemy

httpx==0.24.1
pytest==7.4.0
testcontainers==4.13.2
